+++ lib/sobelow/ci/system.ex
@@ -2,7 +2,10 @@ defmodule Sobelow.CI.System do
   @moduledoc """
   # Command Injection in `System.cmd`
 
-  TODO: <description of vuln>
+  This submodule of the `CI` module checks for Command Injection
+  vulnerabilities through usage of the `System.cmd` function.
+
+  Ensure the the command passed to `System.cmd` is not user-controlled.
 
   `System.cmd` Injection checks can be ignored with the following command: