+++ lib/sobelow/sql/query.ex
@@ -2,7 +2,10 @@ defmodule Sobelow.SQL.Query do
   @moduledoc """
   # SQL Injection in Query
 
-  TODO: <description of vuln>
+  This submodule of the `SQL` module checks for SQL injection
+  vulnerabilities through usage of the `Ecto.Adapters.SQL.query`.
+
+  Ensure that the query is parameterized and not user-controlled.
 
   SQLi Query checks can be ignored with the following command: