+++ lib/sobelow/traversal/file_module.ex
@@ -2,7 +2,10 @@ defmodule Sobelow.Traversal.FileModule do
   @moduledoc """
   # Directory Traversal in `File` function
 
-  TODO: <description of vuln>
+  This submodule checks for directory traversal vulnerabilities in the `File`
+  module.
+
+  Ensure that the path passed to `File` functions is not user-controlled.
 
   File checks can be ignored with the following command: