+++ lib/sobelow/traversal/send_download.ex
@@ -2,7 +2,10 @@ defmodule Sobelow.Traversal.SendDownload do
   @moduledoc """
   # Directory Traversal in `send_download`
 
-  TODO: <description of vuln>
+  This submodule checks for directory traversal vulnerabilities in the
+  `send_download` function of a Phoenix Controller.
+
+  Ensure that the path passed to `send_download` is not user-controlled.
 
   Send Download checks can be ignored with the following command: