+++ lib/sobelow/traversal/send_file.ex
@@ -2,7 +2,10 @@ defmodule Sobelow.Traversal.SendFile do
   @moduledoc """
   # Directory Traversal in `send_file`
 
-  TODO: <description of vuln>
+  This submodule checks for directory traversal vulnerabilities in the
+  `send_file` function.
+
+  Ensure that the path passed to `send_file` is not user-controlled.
 
   Send File checks can be ignored with the following command: